Adding ~evalmanual [FAULTY - need 2nd opinions]
This command isn't finished yet so I will not be submitting a PR for it at the moment. I'm posting it so others may be able to catch any issues I've failed to catch. Problem: Command appears to run fine, however does not submit the edit for the targeted message embed